+++ CHANGES.md
@@ -22,6 +22,9 @@
 * Error - Removal of Resource Causes Failure Instead of Gracefully Informing User.
   [#589](https://github.com/CCI-Tools/cate/issues/589)
 
+* Allow proxy configuration in Cate Desktop
+  [#544](https://github.com/CCI-Tools/cate/issues/544)
+
 
 ## Changes in 2.0.0-dev.8